Refactor: enum클래스 별도 분리
